+++ template/js/checkout.js
@@ -1,4 +1,9 @@
 import './lib/config'
+import './lib/fetch-info'
+
 import '@ecomplus/storefront-app'
 
-import(/* webpackPrefetch: true */ './lib/widgets')
+import './lib/icons'
+import './lib/utils'
+
+import(/* webpackPrefetch: true */ './lib/load-widgets')