+++ template/js/lib/emitter.js
@@ -4,3 +4,13 @@ const emitter = new EventEmitter()
 export default emitter
 
 window.EventEmitter = EventEmitter
+
+const events = {}
+;['on', 'off', 'once'].forEach(method => {
+  events[method] = (ev, fn) => {
+    emitter[method](ev, fn)
+  }
+})
+window._events = events
+
+export { events, emitter }